fix: del unused function, move directory
