+++ projects/portal/src/app/models/wallets/wallet.application.service.ts
@@ -8,7 +8,7 @@ import { UnunifiImportWalletWithPrivateKeyFormDialogComponent } from '../../view
 import { UnunifiKeyFormDialogComponent } from '../../views/dialogs/wallets/ununifi/ununifi-key-form-dialog/ununifi-key-form-dialog.component';
 import { UnunifiSelectCreateImportDialogComponent } from '../../views/dialogs/wallets/ununifi/ununifi-select-create-import-dialog/ununifi-select-create-import-dialog.component';
 import { UnunifiSelectWalletDialogComponent } from '../../views/dialogs/wallets/ununifi/ununifi-select-wallet-dialog/ununifi-select-wallet-dialog.component';
-import { KeplrApplicationService } from './keplr/keplr.application.service';
+import { KeplrApplicationService } from '../keplr/keplr.application.service';
 import { WalletType, StoredWallet } from './wallet.model';
 import { WalletService } from './wallet.service';
 import { Injectable } from '@angular/core';
