+++ projects/portal/src/app/views/dialogs/wallets/keplr/keplr-import-wallet-dialog/keplr-import-wallet-dialog.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialogRef } from '@angular/material/dialog';
 import { cosmosclient } from '@cosmos-client/core';
 import * as crypto from 'crypto';
+import { KeplrApplicationService } from 'projects/portal/src/app/models/keplr/keplr.application.service';
 import { KeyType } from 'projects/portal/src/app/models/keys/key.model';
-import { KeplrApplicationService } from 'projects/portal/src/app/models/wallets/keplr/keplr.application.service';
 import { StoredWallet, WalletType } from 'projects/portal/src/app/models/wallets/wallet.model';
 import { WalletService } from 'projects/portal/src/app/models/wallets/wallet.service';
 import { createCosmosPublicKeyFromUint8Array } from 'projects/portal/src/app/utils/key';
